fix a bug occured when copying the dummy beliefbase.
In the implementation was a bug which gave the copies different ids, so the
angerona framework cant map the beliefbases correctly.

Added a generic unit tests for Beliefbase implementations, the tests for the
concrete Beliefbase implementations like AspBeliefbase should inherit this
test to guarantee that common tests will be run on the beliefbase implementation
The bug describen earlier is handled by those common tests.