implemantation of the Interactive view
the view will be opened by interactive examples. The user can chose the action type an the receiver.

The Agent class is enhances with the attribute type wich shows if the agent is controlled by the User or act on his own (AI).

it remains to make the agent perform his action.